Added size limit
I added a limit to the size of the array, 10000.  When the user is
prompted to enter in an array size, they are told to make it 10000 or
less, if they make it greater they are prompted again, once they enter
in an acceptable number the array is sorted. Also in your
createRandomList function I changed double randomNum = rand() % size; to
int randomNum = rand() % size; It wouldnt build otherwise.